Create jquery.jb.js
Create the file that will contain all the jQuery based logic for the
application. Since it will be using jQuery I have named it using the same
convention. I have implemented the jb namespace and the jb.player
namespace to prevent future naming conflicts.

The jb.player still needs to be implemented.